Refactor & simplified the code of keyboard plugin
